+++ packages/backend/src/index.ts
@@ -18,6 +18,7 @@ backend.add(legacyPlugin('healthcheck', import('./plugins/healthcheck')));
 // auth plugin
 backend.add(import('@backstage/plugin-auth-backend'));
 backend.add(import('@backstage/plugin-auth-backend-module-github-provider'));
+backend.add(import('@internal/plugin-auth-backend-module-gs-provider'));
 // catalog plugin
 backend.add(import('@backstage/plugin-catalog-backend/alpha'));
 backend.add(