+++ plugins/auth-backend-module-gs-provider/src/index.ts
@@ -1,0 +1,8 @@
+/***/
+/**
+ * The gs-provider backend module for the auth plugin.
+ *
+ * @packageDocumentation
+ */
+
+export { authModuleGsProvider as default } from './module';