+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Update Backstage to 1.24.0.
 - Migrate backend to the new backend system.
+- Move GS auth providers configuration to a separate backend module.
 
 ## [0.17.0] - 2024-03-20